4 - PHASE FOUR/backend/pyproject.toml
[project]
name = "evolved-todo-backend"
version = "0.1.0"
description = "FastAPI backend for Evolved Todo web application with SQLModel ORM and Neon PostgreSQL"
readme = "README.md"
authors = [
    { name = "Engineer Abdul Qadir", email = "181554083+EngineerAbdulQadir@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.12",
    "email-validator>=2.3.0",
    "bcrypt<5.0",
    "openai>=1.0.0",
    "mcp>=1.0.0",
    "openai-chatkit>=1.4.1",
    "sse-starlette>=2.0.0",
    "python-dateutil>=2.9.0.post0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "PT", "SIM", "RET", "ARG"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
strict_equality = true

[[tool.mypy.overrides]]
module = "sqlmodel.*"
ignore_missing_imports = true

[build-system]
requires = ["uv_build>=0.8.3,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
